+++ x/xatomic/doc.go
@@ -1,0 +1,3 @@
+// Package xatomic provides atomic utilities. It is an extension to the
+// standard [sync/atomic] package.
+package xatomic