+++ x/xsync/doc.go
@@ -1,0 +1,3 @@
+// Package xsync provides synchronization utilities. It is an extension to the
+// standard [sync] package.
+package xsync